🎨 Zoopla: refactor HTTP library code.
